+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Filter visibility based on catalog.
+
 ## [1.29.5] - 2021-01-11
 
 ### Fixed